+++ ragelo/types/experiment.py
@@ -1,0 +1,692 @@
+"""
+An Experiment is a collection of queries, and manages their retrieved documents, agent answers and their respective evaluations created by Evaluators.
+For most use cases, this class should be the main entry point to interact with RAGElo.
+"""
+
+from __future__ import annotations
+
+import csv
+import json
+import os
+from typing import Literal
+
+from pydantic import Field
+
+from ragelo.logger import logger
+from ragelo.types.evaluables import AgentAnswer, Document
+from ragelo.types.pydantic_models import BaseModel
+from ragelo.types.query import Query
+from ragelo.types.results import (
+    AnswerEvaluatorResult,
+    EloTournamentResult,
+    EvaluatorResult,
+    RetrievalEvaluatorResult,
+)
+
+
+class Experiment(BaseModel):
+    """
+    A class to manage and evaluate an experiment. An experiment is composed of multiple queries and their evaluations.
+    This is the main class for interacting with RAGElo, and manages all the queries to which agents retrieve documents and answer questions.
+    Attributes:
+        experiment_name (str): An unique name for your experiment.
+        queries: (dict[str, Query]): A dictionary with the queries for this experiment. Can also be accessed directly with __getitem__.
+        elo_tournaments: (list[EloTournamentResult]): A list with the results for Elo Tournaments played over this set of queries.
+    Methods:
+        add_query(query, query_id, doc_id, metadata, force): Adds a query (either a Query object or a string) to the current experiment. Returns the id of the added query.
+        add_retrieved_doc(doc, query_id, doc_id, score, agent, force): Adds a retrieved document to a query
+        add_agent_answer(answer, agent, query_id, force): Adds an agent's answer to a query.
+        add_evaluation(evaluation, should_save, force): Adds an evaluation to a retrieved document, an answer or a pairwise game.
+        evaluate_retrieval(metrics, relevance_threshold): Evaluates the retrieval performance of the agents using the specified metrics and the evaluations generated by a `RetrievalEvaluator`. This requires the ir_measures package to be installed.
+        get_qrels(relevance_key, relevance_threshold, output_path, output_format): Gets the qrels (dictionary of relevance) for the queries. Documents with a score above relevance_threshold are considered relevant.  Optionally saves the result to output_path with output_format.
+        get_runs(agents, output_path, output_format): Gets the runs (retrieval results) for the provided agents for all the queries. Optionally saves the result to output_path with output_format.
+        save(): Saves the current state of the queries object to disk as a JSON object to `cache_path`.
+        save_results(result): Saves an evaluation result to the `results_cache_path` file.
+        add_retrieved_docs_from_runfile(run_file_path, corpus, top_k): Adds retrieved documents from a trec-formatted run file to the queries. Will create the Document objects as needed.
+
+        __len__(): Returns the number of queries.
+        __iter__(): Returns an iterator over the queries.
+        __getitem__(key: str) -> Query: Gets a query by its key.
+        keys(): Returns the keys (query_ids) of all queries.
+    """
+
+    experiment_name: str
+    queries: dict[str, Query] = Field(default_factory=dict)
+    elo_tournaments: list[EloTournamentResult] = Field(default_factory=list)
+
+    def __init__(
+        self,
+        save_path: str | None = None,
+        results_save_path: str | None = None,
+        persist_on_disk: bool = True,
+        clear_evaluations: bool = False,
+        csv_path: str | None = None,
+        csv_query_text_col: str = "query",
+        csv_query_id_col: str | None = None,
+        csv_infer_metadata_fields: bool = True,
+        csv_metadata_fields: list[str] | None = None,
+        **data,
+    ):
+        """
+        Initialize the query object with optional data and set up paths and caches.
+
+        Args:
+            save_path (Optional[str]): A JSON file path to persist the queries object on disk. If not set, will create one based on experiment_name. If it already exists, we will try to load the state from that file.
+            results_save_path (Optional[str]): A JSON Lines file path to persist the evaluators result on disk, to avoid re-computing evaluations. If not set, will create one based on experiment_name.
+            persist_on_disk (bool, defaults to True): Whether or not to use save_path and results_save_path to persist results on disk.
+            clear_evaluations (bool, defaults to False): If set to True, will clear all existing evaluations and re-compute as needed.
+            csv_path (Optional[str]): An optional path to a CSV file with the queries to be used in the experiment.
+            csv_query_text_col (str): The column name for query text in the CSV file.
+            csv_query_id_col (Optional[str]): The column name for query ID in the CSV file.
+            csv_infer_metadata_fields (bool, defaults to True): Whether to infer metadata fields from the CSV file.
+            csv_metadata_fields (Optional[list[str]]): The list of metadata fields in the CSV file.
+        Keyword Args:
+            experiment_name (str): The name of your experiment
+
+        The initialization process involves setting up paths and loading queries from different sources:
+        1. If `csv_path` is provided, queries are loaded from the specified CSV file.
+        2. If `csv_path` is not provided but `save_path` exists and points to a valid file, queries are loaded from the cache file.
+        3. If both `csv_path` and `save_path` are provided, the method ensures that the queries loaded from the CSV file match those in the cache file.
+        The method prioritizes loading queries from the CSV file over the cache file. If neither source is available, it relies on the `queries` attribute if it has been set.
+        """
+
+        super().__init__(**data)
+        self.save_cache = persist_on_disk
+        self.cache_path = save_path
+        self.results_cache_path = results_save_path
+
+        if self.cache_path is None:
+            os.makedirs("cache", exist_ok=True)
+            self.cache_path = f"cache/{self.experiment_name}.json"
+
+        if self.results_cache_path is None:
+            self.results_cache_path = self.cache_path.replace(".json", "_results.jsonl")
+
+        # Check where to load the data from. If the cache file exists or the csv_file exists, load from there. If the queries was also set, make sure they match.
+        if csv_path:
+            csv_queries = self._read_queries_csv(
+                csv_path,
+                csv_query_text_col,
+                csv_query_id_col,
+                csv_infer_metadata_fields,
+                csv_metadata_fields,
+            )
+            if len(self.queries) > 0 and set(csv_queries.keys()) != set(
+                self.queries.keys()
+            ):
+                raise ValueError(
+                    "Queries loaded from the CSV file do not match the queries passed"
+                )
+            self.queries = csv_queries
+        elif os.path.isfile(self.cache_path):
+            cache_queries = self._load_cached_queries()
+            if len(self.queries) > 0 and set(cache_queries.keys()) != set(
+                self.queries.keys()
+            ):
+                raise ValueError(
+                    "Queries loaded from the cache file do not match the queries passed"
+                )
+            self.queries = cache_queries
+        if os.path.isfile(self.results_cache_path):
+            self._load_results_from_cache()
+        if clear_evaluations:
+            self.__clear_all_evaluations()
+        self.save()
+
+    def add_query(
+        self,
+        query: Query | str,
+        query_id: str | None = None,
+        metadata: dict | None = None,
+        force: bool = False,
+    ) -> str:
+        """
+        Adds a query to the collection of queries.
+        Args:
+            query (Query | str): The query to be added. It can be an instance of the Query class or a string.
+            query_id (str | None, optional): The identifier for the query. If not provided, a default ID will be generated.
+            metadata (dict | None, optional): Additional metadata for the query. Defaults to None.
+            force (bool, optional): Whether to overwrite the query if it already exists. Defaults to False.
+        """
+
+        if isinstance(query, Query):
+            if query_id is not None and query.qid != query_id:
+                logger.warning(
+                    f"Query ID mismatch. Using ID from query object: {query.qid}"
+                )
+            query_id = query.qid
+            query_obj = query
+        elif query_id is None:
+            query_id = f"query_{len(self.queries) + 1}"
+            logger.warning(
+                f"Query ID not provided. Using default query ID: {query_id}."
+            )
+        else:
+            query_obj = Query(qid=query_id, query=query, metadata=metadata)
+        if query_id in self.queries and not force:
+            logger.info(f"Query {query_id} already exists. Use force=True to overwrite")
+            return query_id
+        if query_id in self.queries and force:
+            logger.info(
+                f"Query {query_id} already exists, but force was set to True. Overwriting."
+            )
+        self.queries[query_id] = query_obj
+        return query_id
+
+    def add_retrieved_doc(
+        self,
+        doc: Document | str,
+        query_id: str | None = None,
+        doc_id: str | None = None,
+        score: float | None = None,
+        agent: str | None = None,
+        force: bool = False,
+    ):
+        """
+        Adds a retrieved document to a query.
+        Args:
+            doc (Document | str): The document to be added. It can be an instance of the Document class or a string.
+            query_id (Optional[str]): The query used to retrieve the document.
+            doc_id (Optional[str]): The identifier for the document. If not provided, a default ID will be generated.
+            score (Optional[float]): The score of the document. Defaults to None.
+            agent (Optional[str]): The agent that retrieved the document. Defaults to None.
+            force (bool): Whether to overwrite the document if it already exists. Defaults to False.
+        """
+        if isinstance(doc, Document):
+            if doc.qid != query_id and query_id is not None:
+                logger.warning(
+                    f"Query ID mismatch. Using ID from document object: {doc.qid}"
+                )
+                query_id = doc.qid
+            if doc.did != doc_id and doc_id is not None:
+                logger.warning(
+                    f"Document ID mismatch. Using ID from document object: {doc.did}"
+                )
+            doc_id = doc.did
+            query_id = doc.qid
+        elif doc_id is None:
+            raise ValueError(
+                "Document was provided as string, but Document ID was not provided"
+            )
+        if query_id is None:
+            raise ValueError(
+                f"ID of the query to which document {doc_id} was retrieved was not provided."
+            )
+        if isinstance(doc, str):
+            doc = Document(qid=query_id, did=doc_id, text=doc)
+        if query_id not in self.queries:
+            raise ValueError(f"Query {query_id} not found in queries")
+        self.queries[query_id].add_retrieved_doc(doc, score, agent, force)
+
+    def add_agent_answer(
+        self,
+        answer: AgentAnswer | str,
+        agent: str | None = None,
+        query_id: str | None = None,
+        force: bool = False,
+    ):
+        if isinstance(answer, AgentAnswer):
+            if answer.qid != query_id and query_id is not None:
+                logger.warning(
+                    f"Query ID mismatch. Using ID from answer object: {answer.qid}"
+                )
+            query_id = answer.qid
+            if answer.agent != agent and agent is not None:
+                logger.warning(
+                    f"Agent mismatch. Using agent from answer object: {answer.agent}"
+                )
+        elif query_id is None:
+            raise ValueError("Query ID not provided")
+        else:
+            if agent is None:
+                agent = f"agent_{len(self.queries[query_id].answers) + 1}"
+                logger.warning(
+                    f"Agent not provided. Using default agent name: {agent}."
+                )
+            if query_id is None:
+                raise ValueError("Query ID not provided")
+            answer = AgentAnswer(qid=query_id, agent=agent, text=answer)
+        if query_id not in self.queries:
+            raise ValueError(f"Query {query_id} not found in queries")
+        self.queries[query_id].add_agent_answer(answer, force=force)
+
+    def add_evaluation(
+        self,
+        evaluation: (
+            RetrievalEvaluatorResult | AnswerEvaluatorResult | EloTournamentResult
+        ),
+        should_save: bool = True,
+        force: bool = False,
+    ):
+        """
+        Adds an evaluation to the queries and optionally saves the result.
+
+        Args:
+            evaluation (RetrievalEvaluatorResult | AnswerEvaluatorResult): The evaluation result to be added.
+            should_save (bool): Flag indicating whether the result should be persist to disk. Defaults to True.
+            force (bool): Whether to overwrite an evaluation if it already exists.
+        """
+        if isinstance(evaluation, EloTournamentResult):
+            self.elo_tournaments.append(evaluation)
+        else:
+            qid = evaluation.qid
+            if qid not in self.queries:
+                raise ValueError(f"Query {qid} not found in queries")
+            self.queries[qid].add_evaluation(evaluation, force=force)
+        if should_save:
+            self.save_results(evaluation)
+
+    def evaluate_retrieval(
+        self,
+        metrics: list[str] = ["Precision@10", "nDCG@10", "Judged@10"],
+        relevance_threshold: int = 0,
+    ) -> dict[str, dict[str, float]]:
+        """
+        Evaluate the retrieval performance of agents using specified metrics.
+        Args:
+            metrics (list[str]): A list of metric names to evaluate (default is ["Precision@10", "nDCG@10", "Judged@10"]).
+        relevance_threshold (int): The threshold above which a document is considered relevant (default is 0).
+        Returns:
+            dict[str, dict[str, float]]: A dictionary where keys are agent names and values are dictionaries of metric scores.
+        Raises:
+            ImportError: If the `ir_measures` package is not installed.
+            ValueError: If a specified metric is not found in the `ir_measures` registry.
+        Notes:
+        - The function uses the `ir_measures` package to calculate the metrics.
+        - If the `rich` package is installed, the results are printed in a formatted table.
+        - If the `rich` package is not installed, the results are printed using plain `print`.
+
+        Example:
+        >>> results = evaluate_retrieval(metrics=["Precision@10", "nDCG@10"], relevance_threshold=1)
+        >>> print(results)
+        {
+            "agent1": {
+                "Precision@10": 0.5,
+                "nDCG@10": 0.6
+            },
+            "agent2": {
+                "Precision@10": 0.4,
+                "nDCG@10": 0.5
+            }
+        }
+        """
+
+        try:
+            import ir_measures
+            from ir_measures import parse_measure
+        except ImportError:
+            raise ImportError(
+                "ir_measures is not installed. Please install it with `pip install ir-measures`"
+            )
+        qrels = self.get_qrels(relevance_threshold=relevance_threshold)
+        runs = self.get_runs()
+        measures = []
+        for metric in metrics:
+            try:
+                measure = parse_measure(metric)
+            except NameError:
+                valid_metrics = list(ir_measures.measures.registry.keys())
+                raise ValueError(
+                    f"Metric {metric} not found. Valid metrics are: {valid_metrics}"
+                )
+            measures.append(measure)
+        metrics = [str(m) for m in measures]
+        results = {}
+        for agent, runs in self.get_runs().items():
+            # Transform the keys of the results back to strings
+            results[agent] = {
+                str(k): v
+                for k, v in ir_measures.calc_aggregate(measures, qrels, runs).items()  # type: ignore
+            }
+
+        key_metric = metrics[0]
+        max_agent_len = max([len(agent) for agent in results.keys()]) + 3
+        max_metric_len = max([len(metric) for metric in metrics])
+        sorted_agents = sorted(
+            results.items(), key=lambda x: x[1][key_metric], reverse=True
+        )
+        try:
+            import rich
+
+            rich.print(f"---[bold cyan] Retrieval Scores [/bold cyan] ---")
+            if relevance_threshold > 0:
+                rich.print(
+                    f"[bold yellow]Relevance threshold: {relevance_threshold}[/bold yellow]"
+                )
+            header = f"[bold magenta]{'Agent Name':<{max_agent_len}}"
+            header += "\t".join([f"{m:<{max_metric_len}}" for m in metrics])
+            header += "[/bold magenta]"
+            rich.print(f"[bold cyan]{header}[/bold cyan]")
+            for agent, scores in sorted_agents:
+                row = f"[bold white]{agent:<{max_agent_len}}[/bold white]"
+                row += "\t".join(
+                    [f"{scores[metric]:<{max_metric_len},.4f}" for metric in metrics]
+                )
+                rich.print(row)
+
+        except ImportError:
+            logger.warning("Rich not installed. Using plain print")
+            print(results)
+        return results
+
+    def get_qrels(
+        self,
+        relevance_key: str | None = "answer",
+        relevance_threshold: int = 0,
+        output_path: str | None = None,
+        output_format: Literal["trec", "json"] = "trec",
+    ) -> dict[str, dict[str, int]]:
+        """
+        Retrieve the qrels (query relevance judgments) for the queries.
+        Args:
+            relevance_key (str | None): The key to use for determining relevance. Defaults to "answer".
+            relevance_threshold (int): The threshold value for relevance. Defaults to 0.
+        Returns:
+            dict[str, dict[str, int]]: A dictionary where each key is a query ID and the value is another dictionary
+                                       mapping document IDs to their relevance scores.
+        """
+
+        qrels = {}
+        for qid, query in self.queries.items():
+            qrels[qid] = query.get_qrels(relevance_key, relevance_threshold)
+        if output_path:
+            with open(output_path, "w") as f:
+                if output_format.lower() == "trec":
+                    for qid, qrel in qrels.items():
+                        for did, rel in qrel.items():
+                            f.write(f"{qid} Q0 {did} {rel}\n")
+                elif output_format.lower() == "json":
+                    json.dump(qrels, f)
+                else:
+                    raise ValueError(
+                        f"Invalid output format for QRELS: {output_format}"
+                        "Valid options are 'trec' and 'json'"
+                    )
+        return qrels
+
+    def get_runs(
+        self,
+        agents: list[str] | None = None,
+        output_path: str | None = None,
+        output_format: str = "trec",
+    ) -> dict[str, dict[str, dict[str, float]]]:
+        """
+        Retrieve runs for specified agents.
+        This method aggregates runs from all queries and organizes them by agent.
+        Args:
+            agents (list[str] | None): A list of agent names to filter the runs. If None, runs for all agents are retrieved.
+            output_path (str | None): The path to save the runs to disk. If None, the runs are not saved.
+            output_format (str): The format to save the runs. Defaults to "trec". If TREC, output_path should be a directory, and each agent will have a separate file.
+        Returns:
+            dict[str, dict[str, dict[str, float]]]: A nested dictionary where the first level keys are agent names,
+            the second level keys are run identifiers, and the third level keys are metric names with their corresponding float values.
+        """
+
+        runs_by_agent = {}
+        for query in self.queries.values():
+            runs = query.get_runs(agents)
+            for agent, runs in runs.items():
+                runs_by_agent[agent] = runs_by_agent.get(agent, {}) | runs
+        if output_path:
+            if output_format.lower() == "trec":
+                if not os.path.isdir(output_path):
+                    os.makedirs(output_path, exist_ok=True)
+                for agent, runs in runs_by_agent.items():
+                    with open(f"{output_path}/{agent}.run", "w") as f:
+                        for qid, docs in runs.items():
+                            sorted_scores = sorted(
+                                docs.items(), key=lambda x: x[1], reverse=True
+                            )
+                            for idx, (did, score) in enumerate(sorted_scores):
+                                f.write(f"{qid} Q0 {did} {idx+1} {score} {agent}\n")
+            elif output_format.lower() == "json":
+                with open(output_path, "w") as f:
+                    json.dump(runs_by_agent, f)
+
+        return runs_by_agent
+
+    def save(self):
+        """
+        Persist the current state of the model to disk if caching is enabled.
+        This method checks if caching is enabled by evaluating the `save_cache` attribute.
+        If caching is enabled, it attempts to write the model's current state to the file
+        specified by `cache_path`. If `cache_path` is not set, a ValueError is raised.
+        Raises:
+            ValueError: If `cache_path` is None and caching is enabled.
+        """
+
+        if not self.save_cache:
+            return
+        if self.cache_path is None:
+            raise ValueError("Cache path not set. Cannot dump queries")
+        with open(self.cache_path, "w") as f:
+            json.dump(self.model_dump(), f)
+
+    def save_results(
+        self,
+        result: EvaluatorResult | EloTournamentResult,
+    ):
+        """
+        Persist the evaluation result to a cache file.
+        This method writes the given evaluation result to a cache file specified by
+        `self.results_cache_path`. The result is serialized to JSON format and appended
+        to the file. The type of the result (either "answer" or "retrieval") is included
+        in the serialized data.
+        Args:
+            result (EvaluatorResult | EloTournamentResult): The evaluation result
+                to be persisted to the cache.
+        Raises:
+            ValueError: If `self.results_cache_path` is not set or if the type of `result`
+                is not recognized.
+        """
+
+        if not self.save_cache:
+            return
+        if self.results_cache_path is None:
+            raise ValueError("Results cache path not set. Cannot dump result")
+        with open(self.results_cache_path, "a+") as f:
+            if isinstance(result, AnswerEvaluatorResult):
+                result_type = "answer"
+            elif isinstance(result, RetrievalEvaluatorResult):
+                result_type = "retrieval"
+            elif isinstance(result, EloTournamentResult):
+                result_type = "elo_tournament"
+            else:
+                raise ValueError(
+                    f"Cannot save evaluation of type {type(result)} to cache"
+                )
+            f.write(json.dumps({result_type: result.model_dump()}) + "\n")
+
+    def add_retrieved_docs_from_runfile(
+        self,
+        run_file_path: str,
+        corpus: dict[str, Document] | dict[str, str],
+        top_k: int | None = None,
+    ):
+        """
+        Adds the retrieved documents from a run file to the queries.
+
+        A run file has, traditionally, the following format:
+            qid, Q0, docid, rank, score, run_name
+        The run name will be considered as the agent that retrieved the documents. The second column is usually ignored.
+        Args:
+            run_file_path (str): Path to the run file.
+            corpus (dict[str, Document] | dict[str, str]): A dictionary with the documents.
+                If a dictionary with just docid and text is provided, we will use it as a corpus.
+                Otherwise, we will assume it's a dictionary with docid as key and Document object as value.
+            top_k (int): The maximum number of documents to add for each query. Assumes that the run file is ordered by rank.
+        """
+        documents_read = set()
+        missing_docs = set()
+        docs_per_query = {}
+        warned_queries = set()
+
+        if isinstance(list(corpus.values())[0], str):
+            corpus = {k: str(v) for k, v in corpus.items()}
+            # corpus = {k: Document(did=k, text=str(v)) for k, v in corpus.items()}
+
+        if not os.path.isfile(run_file_path):
+            raise FileNotFoundError(f"Run file {run_file_path} not found")
+        for line in open(run_file_path):
+            qid, _, did, _, score, agent = line.strip().split()
+            if did not in corpus:
+                missing_docs.add(did)
+                continue
+            doc = corpus[did]
+            if isinstance(doc, str):
+                doc = Document(qid=qid, did=did, text=doc)
+            if qid not in self.queries:
+                if qid not in warned_queries:
+                    warned_queries.add(qid)
+                    logger.warning(f"Query {qid} not found in queries. Skipping")
+                continue
+            if qid not in docs_per_query:
+                docs_per_query[qid] = 0
+            if docs_per_query[qid] < top_k:
+                self.add_retrieved_doc(doc, qid, score=float(score), agent=agent)
+                docs_per_query[qid] += 1
+                documents_read.add(did)
+        if len(missing_docs) > 0:
+            logger.warning(
+                f"Loaded {len(documents_read)} documents. {len(missing_docs)} missing docs"
+            )
+
+    def __clear_all_evaluations(self, should_save: bool = False):
+        logger.warning(f"Clearing all evaluations for {len(self)} queries")
+        doc_eval_count = 0
+        game_eval_count = 0
+        answer_eval_count = 0
+        for query in self.queries.values():
+            for doc in query.retrieved_docs.values():
+                if doc.evaluation is not None:
+                    doc_eval_count += 1
+                doc.evaluation = None
+            for game in query.pairwise_games:
+                if game.evaluation is not None:
+                    game_eval_count += 1
+                game.evaluation = None
+            for answer in query.answers.values():
+                if answer.evaluation is not None:
+                    answer_eval_count += 1
+                answer.evaluation = None
+        logger.info(
+            f"Cleared {doc_eval_count} document evaluations, {game_eval_count} game evaluations, and {answer_eval_count} answer evaluations"
+        )
+        if should_save:
+            if self.results_cache_path and os.path.isfile(self.results_cache_path):
+                with open(self.results_cache_path, "w") as f:
+                    pass
+            self.save()
+
+    def _read_queries_csv(
+        self,
+        csv_path: str,
+        csv_query_text_col: str = "query",
+        csv_query_id_col: str | None = None,
+        csv_infer_metadata_fields: bool = True,
+        csv_metadata_fields: list[str] | None = None,
+    ) -> dict[str, Query]:
+        queries: dict[str, Query] = {}
+        read_queries = set()
+        if not os.path.isfile(csv_path):
+            raise FileNotFoundError(f"CSV file {csv_path} not found")
+        if csv_query_id_col is None:
+            csv_query_id_col = self.__infer_query_id_column(csv_path)
+        with open(csv_path) as f:
+            reader = csv.DictReader(f)
+            headers = reader.fieldnames
+            if headers and csv_infer_metadata_fields:
+                if csv_metadata_fields is None:
+                    csv_metadata_fields = [
+                        x
+                        for x in headers
+                        if x not in [csv_query_id_col, csv_query_text_col]
+                    ]
+                else:
+                    # Use only the metadata fields that actually exist
+                    csv_metadata_fields = [
+                        x for x in csv_metadata_fields if x in headers
+                    ]
+                    if not csv_metadata_fields:
+                        logger.warning(
+                            "No metadata fields found in the csv. Ignoring metadata fields"
+                        )
+                        csv_metadata_fields = None
+            for idx, row in enumerate(reader):
+                qid = row.get(csv_query_id_col) or f"query_{idx}"
+                if qid in read_queries:
+                    logger.warning(f"Query with ID {qid} already read. Skipping")
+                    continue
+                query_text = row[csv_query_text_col].strip()
+                metadata = None
+                if csv_metadata_fields is not None:
+                    metadata = {
+                        k: v for k, v in row.items() if k in csv_metadata_fields
+                    }
+                queries[qid] = Query(qid=qid, query=query_text, metadata=metadata)
+                read_queries.add(qid)
+        return queries
+
+    def _load_cached_queries(self):
+        queries: dict[str, Query] = {}
+        if not self.cache_path:
+            raise ValueError("Cache path not set. Cannot load queries")
+        if not os.path.isfile(self.cache_path):
+            raise FileNotFoundError(f"Cache file {self.cache_path} not found")
+        with open(self.cache_path) as f:
+            queries = {k: Query(**v) for k, v in json.load(f)["queries"].items()}
+        return queries
+
+    def _load_results_from_cache(self):
+        if not self.cache_path:
+            raise ValueError("Cache path not set. Cannot load queries")
+        if self.results_cache_path is None:
+            self.results_cache_path = self.cache_path.replace(".json", "_results.jsonl")
+            with open(self.results_cache_path, "w") as f:
+                pass
+        for line in open(self.results_cache_path):
+            result = json.loads(line)
+            result_type = list(result.keys())[0]
+            if result_type == "answer":
+                result = AnswerEvaluatorResult(**result["answer"])
+                if result.qid not in self.queries:
+                    continue
+            elif result_type == "retrieval":
+                result = RetrievalEvaluatorResult(**result["retrieval"])
+                if result.qid not in self.queries:
+                    continue
+            elif result_type == "elo_tournament":
+                result = EloTournamentResult(**result["elo_tournament"])
+            self.add_evaluation(result, should_save=False)
+
+    def __len__(self):
+        return len(self.queries)
+
+    def __iter__(self):
+        return iter(self.queries.values())
+
+    def __getitem__(self, key: str) -> Query:
+        return self.queries[key]
+
+    @staticmethod
+    def __infer_query_id_column(file_path: str) -> str | None:
+        """Infer the column name with the query id from a CSV file."""
+        possible_qid_columns = [
+            "query_id",
+            "qid",
+            "question_id",
+            "q_id",
+            "queryid",
+            "questionid",
+        ]
+        with open(file_path) as f:
+            reader = csv.DictReader(f)
+            columns = reader.fieldnames
+            if columns is not None:
+                for col in possible_qid_columns:
+                    if col in columns:
+                        return col
+        return None
+
+    def keys(self):
+        return self.queries.keys()
